Admin view - route user/put now included in user/update and legacy removed;
