+++ public/views/_user.html
@@ -591,7 +591,7 @@
 
     function submit(user) {
       
-      fetch(`${document.head.dataset.dir}/api/user/put`, {
+      fetch(`${document.head.dataset.dir}/api/user/update`, {
         method: "POST",
         body: JSON.stringify(user),
         headers: {
